Remove logo from all but README.md.
It seems to interfere with the detection of top-level header title by
vuepress (which renders kudo.dev) cauising the navigation bar to look
ugly.

This is a quick workaround. Hopefully we can restore it once the
embedding of docs on kudo site is implemented in a nice way using
something like GatsbyJS.